feat: fastify caching
