+++ package.json
@@ -41,6 +41,7 @@
     "push-release": "git push --follow-tags origin master"
   },
   "peerDependencies": {
+    "@fastify/caching": "^9.x",
     "@fastify/middie": "^9.x",
     "@fastify/rate-limit": "^10.x",
     "@fastify/reply-from": "^12.x",
@@ -52,6 +53,7 @@
   "devDependencies": {
     "@commitlint/cli": "^19.6.0",
     "@commitlint/config-conventional": "^19.6.0",
+    "@fastify/caching": "^9.0.2",
     "@fastify/middie": "^9.0.2",
     "@fastify/rate-limit": "^10.2.1",
     "@fastify/reply-from": "^12.0.0",
@@ -63,8 +65,8 @@
     "globals": "^15.13.0",
     "husky": "^9.1.7",
     "lint-staged": "^15.2.11",
+    "nock": "^14.0.0-beta.19",
     "prettier": "^3.4.2",
-    "restana": "^4.9.9",
     "standard-version": "^9.5.0"
   },
   "lint-staged": {
